+++ src/steps/create-registries/index.ts
@@ -1,3 +1,4 @@
 export * from './get-base-component-name.js';
 export * from './has-registry.js';
 export * from './pass-component-name-to-base-component.js';
+export * from './update-references.js';
